+++ components/org.wso2.carbon.identity.application.authenticator.oidc/src/main/java/org/wso2/carbon/identity/application/authenticator/oidc/debug/OIDCDebugCallbackHandler.java
@@ -18,8 +18,6 @@
 
 package org.wso2.carbon.identity.application.authenticator.oidc.debug;
 
-import org.apache.commons.lang.StringUtils;
-import org.wso2.carbon.identity.debug.framework.DebugFrameworkConstants;
 import org.wso2.carbon.identity.debug.framework.exception.DebugFrameworkServerException;
 import org.wso2.carbon.identity.debug.framework.extension.DebugCallbackHandler;
 import org.wso2.carbon.identity.debug.framework.model.DebugContext;
@@ -31,9 +29,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 /**
- * OAuth-style debug callback handler for OIDC and OIDC-based protocols (Google, GitHub).
- * Routing is performed by the interceptor using the protocol key stored in the debug session;
- * this handler only processes the callback once routed.
+ * Handles the OIDC authorization code callback for the debug flow.
  */
 public class OIDCDebugCallbackHandler implements DebugCallbackHandler {
 
@@ -54,40 +50,8 @@ public class OIDCDebugCallbackHandler implements DebugCallbackHandler {
     public boolean handleCallback(HttpServletRequest request, HttpServletResponse response,
             Map<String, Object> sessionData) throws DebugFrameworkServerException {
 
-        String code = request.getParameter(OIDCDebugConstants.OIDC_CODE_PARAM);
-        String state = request.getParameter(OIDCDebugConstants.OIDC_STATE_PARAM);
-
-        DebugContext context = buildContext(sessionData);
-        setContextProperties(context, code, state);
-
-        if (response.isCommitted()) {
-            return true;
-        }
-
+        DebugContext context = DebugContext.buildFromMap(sessionData);
         processor.processCallback(request, response, context);
         return true;
-    }
-
-    private DebugContext buildContext(Map<String, Object> sessionData) {
-
-        if (sessionData != null && !sessionData.isEmpty()) {
-            return DebugContext.buildFromMap(sessionData);
-        }
-
-        DebugContext context = new DebugContext();
-        context.setProperty(DebugFrameworkConstants.DEBUG_FLOW_TYPE, DebugFrameworkConstants.FLOW_TYPE_CALLBACK);
-        context.setProperty(DebugFrameworkConstants.DEBUG_CONTEXT_CREATED, DebugFrameworkConstants.TRUE_VALUE);
-        context.setProperty(DebugFrameworkConstants.DEBUG_CREATION_TIMESTAMP, System.currentTimeMillis());
-        return context;
-    }
-
-    private void setContextProperties(DebugContext context, String code, String state) {
-
-        if (StringUtils.isNotBlank(code)) {
-            context.setProperty(DebugFrameworkConstants.DEBUG_PROTOCOL_CODE, code);
-        }
-        if (StringUtils.isNotBlank(state)) {
-            context.setProperty(DebugFrameworkConstants.DEBUG_PROTOCOL_STATE, state);
-        }
     }
 }
