+++ components/org.wso2.carbon.identity.application.authenticator.oidc/src/main/java/org/wso2/carbon/identity/application/authenticator/oidc/debug/client/OAuth2TokenClient.java
@@ -26,10 +26,6 @@ import org.apache.oltu.oauth2.client.request.OAuthClientRequest;
 import org.wso2.carbon.identity.application.authenticator.oidc.CustomURLConnectionClient;
 import org.apache.oltu.oauth2.client.response.OAuthJSONAccessTokenResponse;
 import org.apache.oltu.oauth2.common.message.types.GrantType;
-import org.wso2.carbon.identity.application.authenticator.oidc.debug.OIDCDebugConstants;
-
-import java.util.Locale;
-
 
 /**
  * Responsible for performing OAuth2 token exchanges. Isolates HTTP/network logic from higher-level processors.
@@ -55,13 +51,6 @@ public class OAuth2TokenClient {
     public TokenResponse exchangeCodeForTokens(String authorizationCode, String tokenEndpoint, String clientId,
             String clientSecret, String redirectUri, String codeVerifier, String idpName) {
 
-        // Validate all required parameters.
-        TokenResponse validationError = validateRequiredParameters(authorizationCode, tokenEndpoint, clientId,
-                clientSecret, redirectUri);
-        if (validationError != null) {
-            return validationError;
-        }
-
         try {
             OAuthClientRequest request = buildTokenRequest(tokenEndpoint, clientId, clientSecret, redirectUri,
                     authorizationCode, codeVerifier, idpName);
@@ -78,37 +67,6 @@ public class OAuth2TokenClient {
         } catch (Exception e) {
             return handleTokenExchangeError(e, idpName);
         }
-    }
-
-    /**
-     * Validates that all required parameters are present and non-empty.
-     *
-     * @param authorizationCode The authorization code from the IdP.
-     * @param tokenEndpoint The token endpoint URL of the IdP.
-     * @param clientId The OAuth2 client ID.
-     * @param clientSecret The OAuth2 client secret.
-     * @param redirectUri The redirect URI.
-     * @return TokenResponse with error if validation fails, null if all parameters are valid.
-     */
-    private TokenResponse validateRequiredParameters(String authorizationCode, String tokenEndpoint,
-            String clientId, String clientSecret, String redirectUri) {
-
-        if (StringUtils.isBlank(authorizationCode)) {
-            return TokenResponse.error(OIDCDebugConstants.ERROR_CODE_INVALID_REQUEST, "Authorization code is required");
-        }
-        if (StringUtils.isBlank(tokenEndpoint)) {
-            return TokenResponse.error(OIDCDebugConstants.ERROR_CODE_INVALID_REQUEST, "Token endpoint URL is required");
-        }
-        if (StringUtils.isBlank(clientId)) {
-            return TokenResponse.error(OIDCDebugConstants.ERROR_CODE_INVALID_REQUEST, "Client ID is required");
-        }
-        if (StringUtils.isBlank(clientSecret)) {
-            return TokenResponse.error(OIDCDebugConstants.ERROR_CODE_INVALID_REQUEST, "Client secret is required");
-        }
-        if (StringUtils.isBlank(redirectUri)) {
-            return TokenResponse.error(OIDCDebugConstants.ERROR_CODE_INVALID_REQUEST, "Redirect URI is required");
-        }
-        return null;
     }
 
     /**
@@ -161,134 +119,25 @@ public class OAuth2TokenClient {
      */
     private TokenResponse extractTokenResponse(OAuthJSONAccessTokenResponse oAuthResponse, String idpName) {
 
-        String accessToken = oAuthResponse.getAccessToken();
-        String refreshToken = oAuthResponse.getRefreshToken();
         String tokenType = oAuthResponse.getParam("token_type");
         String idToken = oAuthResponse.getParam("id_token");
 
         if (LOG.isDebugEnabled()) {
-            LOG.debug("Token exchange successful for IdP: " + idpName + ", received access_token and token_type.");
+            LOG.debug("Token exchange successful for IdP: " + idpName + ".");
         }
 
-        return TokenResponse.success(accessToken, idToken, refreshToken, tokenType);
+        return TokenResponse.success(idToken, tokenType);
     }
 
-    /**
-     * Handles errors that occur during token exchange.
-     *
-     * @param e The exception that occurred.
-     * @param idpName The IdP name for logging.
-     * @return TokenResponse with error details.
-     */
     private TokenResponse handleTokenExchangeError(Exception e, String idpName) {
 
         String errorMessage = e.getMessage() != null ? e.getMessage() : e.getClass().getSimpleName();
-        String errorCode = extractErrorCode(e);
-        String errorDescription = buildDetailedErrorDescription(e, errorCode);
 
         if (LOG.isDebugEnabled()) {
-            LOG.debug("Token exchange failed with error code: " + errorCode + ", message: " + errorMessage +
-                    ". IdP: " + idpName);
-        }
-        LOG.error("Token exchange failed for IdP: " + idpName + " - Code: " + errorCode +
-                ", Message: " + errorMessage, e);
-
-        return TokenResponse.error(errorCode, errorDescription);
-    }
-
-    /**
-     * Extracts error code from OAuth exceptions with comprehensive pattern matching.
-     *
-     * @param e The exception to extract error code from.
-     * @return Error code string or generic TOKEN_EXCHANGE_ERROR if code cannot be determined.
-     */
-    private String extractErrorCode(Exception e) {
-
-        if (e == null) {
-            return "TOKEN_EXCHANGE_ERROR";
-        }
-        String exceptionMessage = e.getMessage() != null ? e.getMessage().toLowerCase(Locale.ENGLISH) : "";
-        if (exceptionMessage.contains("invalid_client")) {
-            return "INVALID_CLIENT";
-        } else if (exceptionMessage.contains("invalid_grant")) {
-            return "INVALID_GRANT";
-        } else if (exceptionMessage.contains("unauthorized")) {
-            return "UNAUTHORIZED";
-        } else if (exceptionMessage.contains("invalid_request")) {
-            return OIDCDebugConstants.ERROR_CODE_INVALID_REQUEST;
-        } else if (exceptionMessage.contains("unsupported_grant_type")) {
-            return "UNSUPPORTED_GRANT_TYPE";
-        } else if (exceptionMessage.contains("connection")) {
-            return "CONNECTION_ERROR";
-        } else if (exceptionMessage.contains("timeout")) {
-            return "TIMEOUT_ERROR";
-        } else if (exceptionMessage.contains("ssl") || exceptionMessage.contains("certificate")) {
-            return "SSL_CERTIFICATE_ERROR";
-        } else {
-            return "TOKEN_EXCHANGE_ERROR";
-        }
-    }
-
-    /**
-     * Builds a detailed error description from exception details.
-     * Focuses on actionable information without verbose stack traces.
-     *
-     * @param e The exception that occurred.
-     * @param errorCode The error code extracted from the exception.
-     * @return A detailed error description with troubleshooting hints.
-     */
-    private String buildDetailedErrorDescription(Exception e, String errorCode) {
-
-        StringBuilder details = new StringBuilder();
-        String errorMessage = e != null && e.getMessage() != null ? e.getMessage() : "";
-
-        // Add context-specific troubleshooting hints.
-        switch (errorCode) {
-            case "INVALID_CLIENT":
-                details.append("Client credentials are invalid. Verify that the Client ID and Client Secret ")
-                    .append("are correct in the IdP authenticator configuration.");
-                break;
-            case "INVALID_GRANT":
-                details.append("The authorization code may have expired (usually after 5-10 minutes) ")
-                    .append("or was already used. Start the authentication process again to get a new ")
-                    .append("authorization code.");
-                break;
-            case OIDCDebugConstants.ERROR_CODE_INVALID_REQUEST:
-                if (isLikelyWrongTokenEndpoint(errorMessage)) {
-                    details.append("Configured token endpoint may be incorrect.");
-                } else {
-                    details.append("The token request is malformed. Verify redirect URI and PKCE parameters ")
-                            .append("are configured correctly.");
-                }
-                break;
-            case "UNAUTHORIZED":
-                details.append("The IdP rejected the request. Check that client credentials are correct ")
-                    .append("and the authenticator type matches the IdP's requirements.");
-                break;
-            case "CONNECTION_ERROR":
-                details.append("Cannot connect to the IdP token endpoint. Verify the token endpoint URL ")
-                    .append("is correct and the IdP server is reachable.");
-                break;
-            case "TIMEOUT_ERROR":
-                details.append("The request to the IdP token endpoint timed out. ")
-                    .append("Check if the IdP server is running and network connectivity is available.");
-                break;
-            case "SSL_CERTIFICATE_ERROR":
-                details.append("SSL certificate validation failed. Verify that the IdP's SSL certificate ")
-                    .append("is valid and trusted.");
-                break;
-            default:
-                details.append("An error occurred during token exchange. Check the error code and message for details.");
+            LOG.debug("Token exchange failed for IdP: " + idpName + " - " + errorMessage);
         }
 
-        return details.toString();
-    }
-
-    private boolean isLikelyWrongTokenEndpoint(String errorMessage) {
-
-        String normalizedErrorMessage = errorMessage != null ? errorMessage.toLowerCase(Locale.ENGLISH) : "";
-        return normalizedErrorMessage.contains("missing parameters: access_token")
-                || normalizedErrorMessage.contains("missing parameter: access_token");
+        return TokenResponse.error("TOKEN_EXCHANGE_ERROR", errorMessage);
     }
 
 }
